Remove directions to ignore config file changes
During a Lineage merge any changes should be manually resolved to ensure that
all relevant changes are being applied. The current directions may lead someone
to blindly ignore changes in the Lineage configuration file. Since this file is
rarely changed it should always be manually reviewed if it's a file with
conflicts.