Attempt to make search results more deterministic
I observed an issue where there were duplicates of repositories in the results
of the `get_repo_list()` function. This appears to be caused by an issue with
pagination of the GitHub API on page boundaries. I have added a sort preference
in the hopes of making the search results more deterministic to hopefully
offset this problem.