Add type information for Action metadata to the README
Add information about how inputs are interpreted to the README. Add a default
placeholder to the example table for outputs to describe the type of an output.

In both cases the value will be converted to a string because of how inputs and
outputs are processed. This information will just help set expectations for how
an input is interpreted and what kind of data an output represents.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>